py-publisher/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# py-publisher/main.py: 1
pyzmq == 22.3.0
